add action with drodown to a saperate file
